[fix/IS-8] change tsx icons to svg, refactor authInput
